+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 10.6.1
 - SQL Fix regression since 10.0, change sum/cum factors & credit_attempted/earned columns type from double precision to numeric in Update.fnc.php
 - Fix Grades input not displaying for Teachers in Widget.php
 - Raise minimum PHP version from 5.4.45 to 5.5.9 in INSTALL.md, README.md, composer.json, diagnostic.php & Portal.php
+- Fix close PopTable wrapper <div> in Enrollment.inc.php, General_Info.inc.php, Medical.inc.php & Schedule.inc.php
 
 Changes in 10.6
 ---------------